move enemy editor into team setting
